[onert/test] Rename tool nnpackage_run to onert_run
This commit renames test tool nnpackage_run to onert_run.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
