+++ docs/howto/how-to-use-specific-backend.md
@@ -18,7 +18,7 @@ ONE runtime has many ways to use specific backend during inference
 - Same as `nnfw_set_available_backends`
 - Example
 ```bash
-BACKENDS=cpu ./Product/out/bin/nnpackage_run ...
+BACKENDS=cpu ./Product/out/bin/onert_run ...
 ```
 
 ### 2. OP_BACKEND_[OP_TYPE]
@@ -27,7 +27,7 @@ BACKENDS=cpu ./Product/out/bin/nnpackage_run ...
 - Example
   - Execute `Conv2D` operator on ruy backend and others on cpu backend
 ```bash
-OP_BACKEND_Conv2D=ruy BACKENDS="cpu;ruy" ./Product/out/bin/nnpackage_run ...
+OP_BACKEND_Conv2D=ruy BACKENDS="cpu;ruy" ./Product/out/bin/onert_run ...
 ```
 
 ### 3. OP_BACKEND_MAP
@@ -36,5 +36,5 @@ OP_BACKEND_Conv2D=ruy BACKENDS="cpu;ruy" ./Product/out/bin/nnpackage_run ...
 - Example
   - Execute `operator 10` on `acl_cl` backend and others on `acl_neon` backend
 ```bash
-OP_BACKEND_MAP="10=acl_cl" BACKENDS="acl_neon;acl_cl" ./Product/out/bin/nnpackage_run ...
+OP_BACKEND_MAP="10=acl_cl" BACKENDS="acl_neon;acl_cl" ./Product/out/bin/onert_run ...
 ```
