+++ tests/tools/onert_run/src/args.cc
@@ -56,7 +56,7 @@ std::unordered_map<uint32_t, Json::Value> argArrayToMap(const Json::Value &jsonv
 }
 
 // param shape_str is a form of, e.g., "[1, [2, 3], 3, []]" or "h5"
-void handleShapeJsonParam(nnpkg_run::TensorShapeMap &shape_map, const std::string &shape_str)
+void handleShapeJsonParam(onert_run::TensorShapeMap &shape_map, const std::string &shape_str)
 {
   Json::Value root;
   Json::Reader reader;
@@ -95,7 +95,7 @@ void handleShapeJsonParam(nnpkg_run::TensorShapeMap &shape_map, const std::strin
 
 } // namespace
 
-namespace nnpkg_run
+namespace onert_run
 {
 
 Args::Args(const int argc, char **argv)
@@ -237,7 +237,7 @@ void Args::Initialize(void)
     ("write_report,p", po::value<bool>()->default_value(false)->notifier([&](const auto &v) { _write_report = v; }),
          "Write report\n"
          "{exec}-{nnpkg|modelfile}-{backend}.csv will be generated.\n"
-         "e.g. nnpackage_run-UNIT_Add_000-acl_cl.csv.\n"
+         "e.g. onert_run-UNIT_Add_000-acl_cl.csv.\n"
          "{nnpkg|modelfile} name may be changed to realpath if you use symbolic-link.")
     ("shape_prepare", po::value<std::string>()->default_value("[]")->notifier(process_shape_prepare),
          "Please refer to the description of 'shape_run'")
@@ -286,9 +286,9 @@ void Args::Parse(const int argc, char **argv)
                                             "' or '" + o2 + ".");
     };
 
-    // calling, e.g., "nnpackage_run .. -- shape_prepare .. --shape_run .." should theoretically
+    // calling, e.g., "onert_run .. -- shape_prepare .. --shape_run .." should theoretically
     // work but allowing both options together on command line makes the usage and implemenation
-    // of nnpackage_run too complicated. Therefore let's not allow those option together.
+    // of onert_run too complicated. Therefore let's not allow those option together.
     conflicting_options("shape_prepare", "shape_run");
 
     // Cannot use both single model file and nnpackage at once
@@ -303,7 +303,7 @@ void Args::Parse(const int argc, char **argv)
 
   if (vm.count("help"))
   {
-    std::cout << "nnpackage_run\n\n";
+    std::cout << "onert_run\n\n";
     std::cout << "Usage: " << argv[0] << " path to nnpackage root directory [<options>]\n\n";
     std::cout << _options;
     std::cout << "\n";
@@ -352,4 +352,4 @@ bool Args::shapeParamProvided()
   return provided;
 }
 
-} // end of namespace nnpkg_run
+} // end of namespace onert_run
