+++ tools/pareto_profiler/estimator/brute_force_profiler.py
@@ -8,7 +8,7 @@ from utils import progressbar
 
 if __name__ == "__main__":
     parser = ProfileArgs(
-        prog="brute_force_profiler.py", description="Profiles nnpackage_run using oplist")
+        prog="brute_force_profiler.py", description="Profiles onert_run using oplist")
     # Parse arguments
     args = parser.parse_args()
     modelfile = args.model
